Update with latest changes

+++ src/popover/popover.directive.ts
@@ -5,7 +5,7 @@ import {
 import { PopoverConfig } from './popover.config';
 import { ComponentLoader, ComponentLoaderFactory } from 'ngx-bootstrap/component-loader';
 import { PopoverContainerComponent } from './popover-container.component';
-import { PositioningService } from 'ngx-bootstrap/positioning';
+import { PositioningService, AvailbleBSPositions } from 'ngx-bootstrap/positioning';
 import { timer } from 'rxjs';
 import { parseTriggers, Trigger } from 'ngx-bootstrap/utils';
 
@@ -19,26 +19,25 @@ export class PopoverDirective implements OnInit, OnDestroy {
   /** unique id popover - use for aria-describedby */
   popoverId = id++;
   /** sets disable adaptive position */
-  @Input() adaptivePosition: boolean;
+  @Input() adaptivePosition = true;
   /**
    * Content to be displayed as popover.
    */
-  /* tslint:disable-next-line: no-any */
-  @Input() popover: string | TemplateRef<any>;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  @Input() popover?: string | TemplateRef<any>;
   /**
    * Context to be used if popover is a template.
    */
-  /* tslint:disable-next-line: no-any */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   @Input() popoverContext: any;
   /**
    * Title of a popover.
    */
-  @Input() popoverTitle: string;
+  @Input() popoverTitle?: string;
   /**
    * Placement of a popover. Accepts: "top", "bottom", "left", "right"
    */
-  @Input() placement: 'top' | 'bottom' | 'left' | 'right' | 'auto' | 'top left' | 'top right' |
-    'right top' | 'right bottom' | 'bottom right' | 'bottom left' | 'left bottom' | 'left top';
+  @Input() placement: AvailbleBSPositions = 'top';
   /**
    * Close popover on outside click
    */
@@ -47,11 +46,11 @@ export class PopoverDirective implements OnInit, OnDestroy {
    * Specifies events that should trigger. Supports a space separated list of
    * event names.
    */
-  @Input() triggers: string;
+  @Input() triggers = 'click';
   /**
    * A selector specifying the element the popover should be appended to.
    */
-  @Input() container: string;
+  @Input() container?: string;
 
   /**
    * Css class for popover container
@@ -77,26 +76,24 @@ export class PopoverDirective implements OnInit, OnDestroy {
   /**
    * Delay before showing the tooltip
    */
-  @Input() delay: number;
+  @Input() delay = 0;
 
   /**
    * Emits an event when the popover is shown
    */
-  /* tslint:disable-next-line: no-any */
-  @Output() onShown: EventEmitter<any>;
+  @Output() onShown: EventEmitter<unknown>;
   /**
    * Emits an event when the popover is hidden
    */
-  /* tslint:disable-next-line: no-any */
-  @Output() onHidden: EventEmitter<any>;
+  @Output() onHidden: EventEmitter<unknown>;
 
-  protected _popoverCancelShowFn: Function;
+  protected _popoverCancelShowFn?: () => void;
 
-  protected _delayTimeoutId: number | any;
+  protected _delayTimeoutId?: number;
 
   private _popover: ComponentLoader<PopoverContainerComponent>;
   private _isInited = false;
-  private _ariaDescribedby: string;
+  private _ariaDescribedby?: string;
 
   constructor(
     _config: PopoverConfig,
@@ -136,9 +133,11 @@ export class PopoverDirective implements OnInit, OnDestroy {
    * set id for the popover
    */
   setAriaDescribedBy(): void {
-    this._ariaDescribedby = this.isOpen ? `ngx-popover-${this.popoverId}` : null;
+    this._ariaDescribedby = this.isOpen ? `ngx-popover-${this.popoverId}` : void 0;
     if (this._ariaDescribedby) {
-      this._popover.instance.popoverId = this._ariaDescribedby;
+      if (this._popover.instance) {
+        this._popover.instance.popoverId = this._ariaDescribedby;
+      }
       this._renderer.setAttribute(this._elementRef.nativeElement, 'aria-describedby', this._ariaDescribedby);
     } else {
       this._renderer.removeAttribute(this._elementRef.nativeElement, 'aria-describedby');
@@ -182,7 +181,7 @@ export class PopoverDirective implements OnInit, OnDestroy {
           containerClass: this.containerClass
         });
 
-      if (!this.adaptivePosition) {
+      if (!this.adaptivePosition && this._popover._componentRef) {
         this._positionService.calcPosition();
         this._positionService.deletePositionElement(this._popover._componentRef.location);
       }
@@ -206,6 +205,10 @@ export class PopoverDirective implements OnInit, OnDestroy {
       if (this.triggers) {
         parseTriggers(this.triggers)
           .forEach((trigger: Trigger) => {
+            if (!trigger.close) {
+              return;
+            }
+
             this._popoverCancelShowFn = this._renderer.listen(
               this._elementRef.nativeElement,
               trigger.close,
